Pushing a dummy commit to rerun check in PR
Added an extra blank line to tile.json in order to force the PR on Jon's repo to rerun. One of the actions failed to validate gradle wrapper. I can't rerun checks on PRs for some reason.
This is a reoccurring issue it seems like, but anyway, the PR is done and ready to be merged